pretty much working bin

+++ RIA/src/main/java/controllers/RemoveUser.java
@@ -2,6 +2,7 @@ package controllers;
 
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.text.ParseException;
 import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -12,7 +13,9 @@ import javax.servlet.http.HttpSession;
 
 import dao.RelationshipsDAO;
 import dao.UserDAO;
+import dao.GroupDAO;
 import beans.User;
+import beans.Group;
 import utils.ConnectionHandler;
 
 @WebServlet("/RemoveUser")
@@ -59,14 +62,35 @@ public class RemoveUser extends HttpServlet {
 		
 		RelationshipsDAO relDao = new RelationshipsDAO(connection);
 		UserDAO uDao = new UserDAO(connection);
+		GroupDAO gDao = new GroupDAO(connection);
+		Group group = null;
+		int participants_num = -1;
 		
 		try {
+			group = gDao.getGroup(groupId);
+			participants_num = uDao.getGroupParticipants(groupId).size();
+		} catch (SQLException | ParseException e) {
+			response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+			response.getWriter().println("Not possible to get group");
+			return;
+		}
+		
+		// check if the minimum number of participants is still satisfied
+		if (participants_num != -1 && participants_num - 1 < group.getMinParticipants()) {
+			response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+			response.getWriter().println("Too few participants would remain");
+			return;		
+		}
+		
+		try {
+			// check if the user is the group creator
 			User creator = uDao.getCreator(groupId);
 			if (creator.equals(u)) {
 				relDao.removeUser(userId, groupId);
 			} else {
 				response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-				response.getWriter().println("The user is not allowed to perform that action");
+				response.getWriter().println("You're not allowed to perform this action");
+				return;
 			}
 		} catch (SQLException e) {
 			response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
